deploy to docs from "reypte" branch
When we make a commit to the retype branch, we want to deploy the
documentation.

Absolutely. when going live on ddnexus/pagy we would want any commits on
on the master branch to update the documentation site. Full details are
explained here: https://github.com/ddnexus/pagy/pull/380 on how
to configure it all.